+++ test/e2e/helpers/page.js
@@ -6,6 +6,7 @@ export default class Page {
     this.labels = Selector('label');
     this.inputs = this.labels.find('input');
     this.errors = Selector('p.error');
+    this.errorsValues = this.errors.find('.error-value');
     this.submitButton = this.forms.find('[type="submit"]');
     this.resetButton = this.forms.find('[type="reset"]');
     this.slideButton = Selector('button.slide-button');