+++ changelog.md
@@ -1,7 +1,8 @@
 # Svelidation changelog
 
-# 1.2.1
-* rename `typeCheck` method to `type`
+# 1.2.0
+* rename `typeCheck` method and errorsCode to `type`
+* added `useCustomErrorsStore` option
 
 # 1.1.1
 * add number own required rule